requirements.txt
streamlit
transformers
torch
faiss-cpu
sentence-transformers
beautifulsoup4
requests
PyPDF2
python-docx
networkx
plotly
pandas
numpy
scikit-learn
chromadb
langchain
langchain-community
streamlit-option-menu
streamlit-chat
streamlit-agraph
pillow
huggingface_hub
python-dotenv